casting test data to correct type

+++ test/core/table/BufferedTable.cpp
@@ -34,7 +34,7 @@ TEST(BufferedTable, AllGatherv) {
      */
     const size_t irank_empty = mpi::nrank()==1 ? ~0ul: 1ul;
     auto get_nrow = [irank_empty](size_t irank){return irank==irank_empty ? 0ul : hashing::in_range(irank, 3, 10);};
-    auto get_value = [](size_t irank, size_t irow){return hashing::in_range(irow * (irank+1), 0, 100);};
+    auto get_value = [](size_t irank, size_t irow){return int(hashing::in_range(irow * (irank+1), 0, 100));};
     const size_t nrow_local = get_nrow(mpi::irank());
     size_t nrow_global = 0ul;
     for (size_t irank=0ul; irank<mpi::nrank(); ++irank) nrow_global+=get_nrow(irank);
@@ -68,7 +68,7 @@ TEST(BufferedTable, Gatherv) {
      */
     const size_t irank_empty = mpi::nrank()==1 ? ~0ul: 1ul;
     auto get_nrow = [irank_empty](size_t irank){return irank==irank_empty ? 0ul : hashing::in_range(irank, 3, 10);};
-    auto get_value = [](size_t irank, size_t irow){return hashing::in_range(irow * (irank+1), 0, 100);};
+    auto get_value = [](size_t irank, size_t irow){return int(hashing::in_range(irow * (irank+1), 0, 100));};
     const size_t nrow_local = get_nrow(mpi::irank());
     size_t nrow_global = 0ul;
     for (size_t irank=0ul; irank<mpi::nrank(); ++irank) nrow_global+=get_nrow(irank);
